requirements-free.txt
# 100% 무료 RAG 시스템 패키지

# Google Gemini API (무료!)
google-generativeai>=0.3.0

# Chroma DB (로컬 벡터 DB, 무료!)
chromadb>=0.4.0

# PDF 파일 읽기
PyPDF2>=3.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# FastAPI (로컬 API 서버)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# 추가 필수 패키지
numpy>=1.24.0
